updated nav bar logic
