+++ front-end/src/App.js
@@ -1,7 +1,9 @@
 import { React, useState } from "react";
 import { Routes, Route } from "react-router-dom";
 
-import NavBar from "./Components/NavBar";
+import NavBarAdmin from "./Components/NavBar/NavBarAdmin";
+import NavBarLoggedIn from "./Components/NavBar/NavBarLoggedIn";
+import NavBarLoggedOut from "./Components/NavBar/NavBarLoggedOut";
 import Team from "./Pages/Team";
 import YourExercise from "./Pages/YourExercise";
 import Profile from "./Pages/Profile";
@@ -12,6 +14,7 @@ import Challenges from "./Pages/Challenges";
 
 const App = () => {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const [isAdmin, setIsAdmin] = useState(false);
   const [userId, setUserId] = useState("");
   const [userInfo, setUserInfo] = useState({});
 
@@ -29,7 +32,11 @@ const App = () => {
 
   return (
     <div>
-      <NavBar isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn} />
+      {isLoggedIn === false && <NavBarLoggedOut />}
+
+      {isLoggedIn === true && <NavBarLoggedIn />}
+
+      {isLoggedIn === true && isAdmin && <NavBarAdmin />}
 
       <div>
         {isLoggedIn === true && (
