+++ front-end/src/Pages/NotFound.js
@@ -1,0 +1,10 @@
+import { React } from "react";
+
+export default function NotFound() {
+  return (
+    <main style={{ padding: "20px" }}>
+      <h2>Whoops!</h2>
+      This is the 404 page.
+    </main>
+  );
+}
